Update README.md
Nexus Bank is a financial institution that aims to provide exceptional banking services tailored to the specific needs and goals of its clients. They offer a wide range of banking solutions and prioritize delivering high-quality service, transparency, and honesty. However, during a recent board meeting, the directors expressed dissatisfaction with the current state of affairs and recognized the need to optimize operations and improve efficiency. As a Data Scientist, the Director of Nexus Bank has reached out for assistance in leveraging data to gain insights into the bank and enhance its performance. The board wants to analyze customer behavior patterns and trends, particularly focusing on the impact of customer demographics on attitudes towards defaulting, in order to anticipate future customer behavior and determine the likelihood of customer deposits.